int --> double

+++ src/scr/DataPoint.java
@@ -7,11 +7,11 @@ package scr;
 //passare i dati facilmente al classificatore
 public class DataPoint {
     public double[] features;  // track[8], track[9], track[10], trackPos, angle, speed
-    public int steering;       // -1, 0, 1
-    public int accelerate;     // 0 o 1
-    public int brake;          // 0 o 1
+    public double steering;       // -1, 0, 1
+    public double accelerate;     // 0 o 1
+    public double brake;          // 0 o 1
 
-    public DataPoint(double[] features, int steering, int accelerate, int brake) {
+    public DataPoint(double[] features, double steering, double accelerate, double brake) {
         this.features = features;
         this.steering = steering;
         this.accelerate = accelerate;
